postgres specific types

+++ macros/casts/number_field.sql
@@ -1,4 +1,4 @@
-{%- macro number_field(field, cast_type='INT64', case_null=None) -%}
+{%- macro number_field(field, cast_type='integer', case_null=None) -%}
 coalesce(
     {{ safe_cast(field, cast_type) }},
     {{ case_null if case_null is not none else 'null' }}
